implements posting and retrieving of single tracks and tests it (tests fail when server is unavailable)

+++ HikingApp/app/src/main/java/ch/epfl/sweng/team7/network/DatabaseClient.java
@@ -55,5 +55,5 @@ public interface DatabaseClient {
      * @throws DatabaseClientException in case the data could not be
      * retrieved for any reason external to the application (network failure, etc.)
      */
-    int postTrack(ch.epfl.sweng.team7.database.TrackData track) throws DatabaseClientException;
+    long postTrack(ch.epfl.sweng.team7.database.TrackData track) throws DatabaseClientException;
 }
